refactor(point): 애그리거트 루트 단위로 레포지토리 병합
